+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0 - 5th March, 2019
+* Implemented document link provider for `zscriptinsert` command
+* Implemented [workspace symbol providers](https://code.visualstudio.com/docs/editor/editingevolved#_open-symbol-by-name)
+* Implemented folding ranges.
+
 ## 1.1.0 - 4th March, 2019
 * Implement Color Provider
 * Fix command (`IGet`, `ISet`) arg type